refactor: add src/shared for bot+web auth, permissions, player state, websocket
- Move implementations to src/shared; web keeps thin re-exports for @/ paths
- Add src/lib/webPrisma.ts and bot-api-origin.ts (web server re-exports)
- Bot handlers and server import from ../../shared (no src/web from bot)
- Web tsconfig: include ../shared, path @/shared/*

Made-with: Cursor
