+++ src/botApi/handlers/guildList.ts
@@ -1,7 +1,7 @@
-import { auth } from "../../web/auth-node.js"
+import { auth } from "../../shared/auth-node.js"
 import { fetchDiscordUserGuilds } from "../../util/discordUserGuilds.js"
 import { DISCORD_BOT_INVITE_PERMISSIONS } from "../../util/discordBotPermissions.js"
-import { getAuthenticatedSession } from "../../web/lib/api-auth.js"
+import { getAuthenticatedSession } from "../../shared/api-auth.js"
 import { tryGetBotClient } from "../../lib/botClientRegistry.js"
 import type { ApiResponse } from "../../types/index.js"
 import type { GuildListResponse } from "../../types/web.js"
