+++ src/shared/auth-node.ts
@@ -1,0 +1,16 @@
+import { betterAuth } from "better-auth"
+import { prismaAdapter } from "better-auth/adapters/prisma"
+import { betterAuthBaseConfig } from "./auth-base-config.js"
+import { getWebPrismaClient } from "../lib/webPrisma.js"
+
+/**
+ * Better Auth for the **bot process** (Express `/api/guilds/*`, WebSocket upgrade).
+ * Same DB and secrets as Next; no `better-auth/next-js` / `next` dependency.
+ * @see {@link ./auth.ts} for the Next.js dashboard instance.
+ */
+export const auth = betterAuth({
+    ...betterAuthBaseConfig,
+    database: prismaAdapter(getWebPrismaClient(), { provider: "postgresql" }),
+})
+
+export type BetterAuthSession = typeof auth.$Infer.Session
